Restructure SnippetPreview and move SnippetEditor helpers to helper dir
