+++ packages/search-metadata-previews/src/SnippetPreview/FixedWidthContainer.js
