+++ packages/search-metadata-previews/src/SnippetPreview/ProgressBar.js
