+++ packages/search-metadata-previews/src/SnippetPreview/SnippetPreview.js
@@ -13,8 +13,8 @@ import { __ } from "@wordpress/i18n";
 // Internal dependencies.
 import FixedWidthContainer from "./FixedWidthContainer";
 import ScreenReaderText from "yoast-components/a11y/ScreenReaderText";
-import { DEFAULT_MODE, MODE_DESKTOP, MODE_MOBILE, MODES } from "../constants";
-import { angleLeft, angleRight } from "../../SnippetEditor/components/Shared";
+import { DEFAULT_MODE, MODE_DESKTOP, MODE_MOBILE, MODES } from "./constants";
+import { angleLeft, angleRight } from "../SnippetEditor/components/Shared";
 import { colors } from "@yoast/components/style-guide";
 import { getRtlStyle } from "@yoast/components";
