--- packages/search-metadata-previews/src/SnippetPreview/index.js
@@ -1,3 +1,0 @@
-export { default as SnippetPreview } from "./components/SnippetPreview";
-export { default as FixedWidthContainer } from "./components/FixedWidthContainer";
-export { default as constants } from "./constants";
