+++ packages/search-metadata-previews/src/helpers/selection.js
