+++ packages/search-metadata-previews/src/helpers/serialization.js
