+++ packages/search-metadata-previews/src/index.js
@@ -1,2 +1,8 @@
+export { default as constants } from "./SnippetPreview/constants"
+export { default as ProgressBar } from "./SnippetPreview/ProgressBar";
+export * from "./SnippetPreview/SnippetPreview";
+export { default as HelpTextWrapper} from "./SnippetPreview/HelpTextWrapper";
+export { default as FixedWidthContainer} from "./SnippetPreview/FixedWidthContainer"
+
 export * from "./SnippetEditor";
-export * from "./SnippetPreview";
+
